project4
 For this project, we will utilize both semantic and episodic memories. You will be printing a
 series of food items in order, ascending by calorie count. Then, you will use episodic memory to
 print the lowest calorie item again.
 
Your task: Print the items in order of calorie count.
The Semantic Memory contains food and calorie information. The working memory contains the items you have in your fridge. Your task is to propose an operator for each item in the fridge and 'apply' it by printing 'Eating *food-Name-here*'. Break impasses in the substate that is created,
utilizing the Semantic Memory to pick the item with the lowest calorie count. Also be sure to remove and update <s> ^selected each time.
Once you have printed all the items, use episodic memory to print the first item that was printed out.